+++ sources/ft_game/ft_wall_texture_animated.c
@@ -6,20 +6,20 @@
 /*   By: tlassere <tlassere@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/30 19:23:20 by tlassere          #+#    #+#             */
-/*   Updated: 2024/05/01 17:24:44 by tlassere         ###   ########.fr       */
+/*   Updated: 2024/05/02 15:13:54 by tlassere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_cub.h"
 
-static uint8_t	*ft_get_pos_x(mlx_texture_t *texture, float pos_x, int frame)
+static uint8_t	*ft_get_pos_x(mlx_texture_t *texture, double pos_x, int frame)
 {
 	return (texture->pixels + ((size_t)(SIZE_DOOR * pos_x
 			+ (SIZE_DOOR * frame)) * sizeof(int)));
 }
 
 void	ft_print_line_animated(t_data *data, mlx_texture_t *texture,
-	t_data_wall wall, float pos_x)
+	t_data_wall wall, double pos_x)
 {
 	t_texture_pos	texture_pos;
 	size_t			texture_width;
